Proyecto_Final/requirements.txt
Django
mysqlclient
djangorestframework
django-debug-toolbar